refactor skill progress record table
