+++ pages/equines/[equineId]/training-programmes/[trainingProgrammeId].tsx
@@ -16,9 +16,10 @@ import {
 	MenuItem,
 } from "@mui/material";
 import LoadingSpinner from "../../../../components/LoadingSpinner";
+import ArrowRightIcon from "@mui/icons-material/ArrowRight";
 import ArrowLeftIcon from "@mui/icons-material/ArrowLeft";
-import KeyboardArrowRightIcon from "@mui/icons-material/KeyboardArrowRight";
 import TrainingProgrammeLog from "../../../../components/pages/equines/TrainingProgrammeLog";
+import TrainingProgrammeSkills from "../../../../components/pages/equines/TrainingProgrammeSkills";
 
 const TrainingProgrammePage = () => {
 	const router = useRouter();
@@ -27,78 +28,13 @@ const TrainingProgrammePage = () => {
 	>(undefined);
 	const { fetchingData, trainingProgramme, error } =
 		useTrainingProgramme(trainingProgrammeId);
-	const [trainingProgrammeSkills, setTrainingProgrammeSkills] = useState<
-		SkillProgressRecord[] | undefined
-	>(undefined);
-	const [skillsFilter, setSkillsFilter] = useState("all");
 
 	useEffect(() => {
 		if (router.isReady) {
 			settrainingProgrammeId(router.query["trainingProgrammeId"]);
-			setTrainingProgrammeSkills(trainingProgramme?.skillProgressRecords);
 		}
 	}, [router.isReady]);
 
-	const mapProgressCodeToOptions = () => {
-		return Object.keys(ProgressCode)
-			.filter((v) => isNaN(Number(v)))
-			.map((progressCode, i) => {
-				return (
-					<MenuItem key={i} value={progressCode}>
-						{progressCode}
-					</MenuItem>
-				);
-			});
-	};
-
-	const handleSkillsFilterChange = (event) => {
-		setSkillsFilter(event.target.value);
-
-		event.target.value === "all"
-			? setTrainingProgrammeSkills(trainingProgramme?.skillProgressRecords)
-			: setTrainingProgrammeSkills(
-					trainingProgramme?.skillProgressRecords.filter(
-						(skillProgressRecord) =>
-							skillProgressRecord.progressCode.string === event.target.value
-					)
-			  );
-	};
-
-	const mapSkillProgressRecords = () => {
-		const skillProgessRecords =
-			skillsFilter === "all"
-				? trainingProgramme?.skillProgressRecords
-				: trainingProgramme?.skillProgressRecords.filter(
-						(spr) => spr.progressCode.string === skillsFilter
-				  );
-
-		return skillProgessRecords?.map((skillProgessRecord, i) => {
-			return (
-				<Paper key={skillProgessRecord.id}>
-					<Box
-						py={1}
-						px={2}
-						mb={2}
-						sx={{
-							display: "flex",
-							justifyContent: "space-between",
-							alignItems: "center",
-						}}
-					>
-						<div>
-							<Typography fontWeight={600}>
-								{skillProgessRecord.skill.name}
-							</Typography>
-							<Typography mt={1}>
-								{skillProgessRecord.progressCode.string}
-							</Typography>
-						</div>
-						<KeyboardArrowRightIcon fontSize="large" />
-					</Box>
-				</Paper>
-			);
-		});
-	};
 
 	if (error) {
 		return (
@@ -130,37 +66,9 @@ const TrainingProgrammePage = () => {
 					<ArrowLeftIcon /> Back
 				</Link>
 			</Breadcrumbs>
-			<Box
-				mt={4}
-				sx={{
-					display: "flex",
-					justifyContent: "space-between",
-					alignItems: "flex-end",
-				}}
-			>
-				<Typography variant="h5" color="gray">
-					Skills
-				</Typography>
-				<FormControl size="small" variant="outlined" sx={{ minWidth: "150px" }}>
-					<Select
-						id="progressCode"
-						value={skillsFilter}
-						onChange={handleSkillsFilterChange}
-					>
-						<MenuItem value="all">Show All</MenuItem>
-						{mapProgressCodeToOptions()}
-					</Select>
-				</FormControl>
-			</Box>
-			<hr style={{ margin: "16px 0" }} />
-			<Box>
-				{trainingProgramme && mapSkillProgressRecords()}
-				{trainingProgrammeSkills && trainingProgrammeSkills.length === 0 && (
-					<Typography>
-						<em>No skills marked as "{skillsFilter}"</em>
-					</Typography>
-				)}
-			</Box>
+			<TrainingProgrammeSkills
+				skillProgressRecords={trainingProgramme?.skillProgressRecords}
+			/>
 			<TrainingProgrammeLog
 				skillTrainingSessions={trainingProgramme?.skillTrainingSessions}
 			/>
